devices/ca: extract the acq_timestamp monitor and trigger into reusable mixins
CaAcqTimestampReadable and CaTriggerable keep their public surface and
private state names (_last_acq, _shot_queue, _monitoring); the bodies move
verbatim to shot_monitor.py so the namespace device (#807 phase 1) can be
shot-aware without a second implementation.
